fix for error when terminating an already complete proces #40

+++ gnomecast.py
@@ -161,7 +161,7 @@ class Transcoder(object):
   
   def destroy(self):
     self.cast.media_controller.stop()
-    if self.p:
+    if self.p and self.p.poll() is None:
       self.p.terminate()
     if self.trans_fn and os.path.isfile(self.trans_fn):
       os.remove(self.trans_fn)
